Validar "Usuario Normal" o "Administrador"

+++ app/Http/Controllers/UsersController.php
@@ -16,16 +16,107 @@ class UsersController extends Controller
         $this->middleware('auth');
     }
 
+
+    public function validateAdmin()
+    {
+        if(auth()->user()->role != "Administrador"){
+
+            return false;
+        }
+    }
+
+
     public function indexUsers()
     {
+        $isAdmin = $this->validateAdmin();
+        if (!$isAdmin) return view('home');
+
         $users = User::all();
 
         return view('usersViews.showUsers')->with('users', $users);
     }
 
 
+    public function upload(Request $request)
+    {
+        $isAdmin = $this->validateAdmin();
+        if (!$isAdmin) return view('home');
+
+        if($request->hasFile("url")){
+
+            $file = $request->file("url");
+            $name = $request->file("url")->getClientOriginalName();
+            $path = public_path("storage/".$name);
+
+            copy($file, $path);
+
+            $user = DB::table('users')
+                        ->where('email','=', $request->userEmail)
+                        ->get();
+
+            $idUser = $user[0]->id;        
+            $nameUser =$user[0]->name;
+
+            DB::table('files')->insert([
+                'idUser' => $idUser,
+                'nameUser' => $nameUser,
+                'file' => $name
+            ]);
+
+            return redirect('ShowUsers')->with('FileUploaded', 'OK');
+
+        }
+
+        return redirect('ShowUsers')->with('ErrorUploaded', 'OK');
+    }
+
+
+    public function createIndex()
+    {
+        $isAdmin = $this->validateAdmin();
+        if (!$isAdmin) return view('home');
+
+        return view('usersViews.createUser');
+    }
+
+
+    public function createUser(Request $request)
+    {
+
+        $isAdmin = $this->validateAdmin();
+        if (!$isAdmin) return view('home');
+
+        $validate = request()->validate([
+
+            'name' => ['required', 'string', 'max:255'],
+            'email' => ['required', 'string', 'email', 'max:255', 'unique:users'],
+            'password' => ['required', 'string', 'min:8', 'confirmed'],
+
+        ]);
+
+        User::create([
+            'name' => $request->name,
+            'email' => $request->email,
+            'role' => $request->rol,
+            'password' => Hash::make($request->password),
+        ]);
+
+        return redirect('ShowUsers')->with('UserCreated', 'OK');
+    }
+
+
     public function editUser($id)
     {
+
+        $isAdmin = $this->validateAdmin();
+        if (!$isAdmin) return view('home');
+
+        $isAdmin = $this->validateAdmin();
+        
+        if(!$isAdmin){
+
+            return view('home');
+        }
 
         $users = User::all();
         $user = User::find($id);
@@ -66,9 +157,13 @@ class UsersController extends Controller
         return redirect('ShowUsers')->with('UserEdited', 'OK');
 
     }
+    
 
     public function destroy($id)
     {
+        $isAdmin = $this->validateAdmin();
+        if (!$isAdmin) return view('home');
+
         $user = User::find($id);
                 
         User::destroy($id);
@@ -76,59 +171,6 @@ class UsersController extends Controller
         return redirect('ShowUsers')->with('UserDeleted', 'OK');
     }
 
-    public function upload(Request $request)
-    {
-        if($request->hasFile("url")){
 
-            $file = $request->file("url");
-            $name = $request->file("url")->getClientOriginalName();
-            $path = public_path("storage/".$name);
 
-            copy($file, $path);
-
-            $user = DB::table('users')
-                        ->where('email','=', $request->userEmail)
-                        ->get();
-
-            $idUser = $user[0]->id;        
-            $nameUser =$user[0]->name;
-
-            DB::table('files')->insert([
-                'idUser' => $idUser,
-                'nameUser' => $nameUser,
-                'file' => $name
-            ]);
-
-            return redirect('ShowUsers')->with('FileUploaded', 'OK');
-
-        }
-
-        return redirect('ShowUsers')->with('ErrorUploaded', 'OK');
-    }
-
-    public function createIndex()
-    {
-        return view('usersViews.createUser');
-    }
-
-    public function createUser(Request $request)
-    {
-
-        $validate = request()->validate([
-
-            'name' => ['required', 'string', 'max:255'],
-            'email' => ['required', 'string', 'email', 'max:255', 'unique:users'],
-            'password' => ['required', 'string', 'min:8', 'confirmed'],
-
-        ]);
-
-        User::create([
-            'name' => $request->name,
-            'email' => $request->email,
-            'role' => $request->rol,
-            'password' => Hash::make($request->password),
-        ]);
-
-        return redirect('ShowUsers')->with('UserCreated', 'OK');
-    }
 }
